- moved the domain layer to a separate module
