+++ app/src/main/java/com/example/topmovies/models/mapper/MovieResponseMapper.kt
@@ -1,6 +1,6 @@
 package com.example.topmovies.models.mapper
 
-import com.example.topmovies.models.domain.Movie
+import com.example.domain.models.Movie
 import com.example.topmovies.models.response.MovieResponse
 
 object MovieResponseMapper : Mapper<MovieResponse, Movie> {
