+++ app/src/main/java/com/example/topmovies/viewmodel/BaseViewModel.kt
@@ -3,7 +3,7 @@ package com.example.topmovies.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.topmovies.utils.Error
+import com.example.domain.utils.Error
 
 abstract class BaseViewModel : ViewModel() {
