+++ domain/src/main/java/com/example/domain/models/Movie.kt
@@ -1,4 +1,4 @@
-package com.example.topmovies.models.domain
+package com.example.domain.models
 
 data class Movie(
     val id: String,
