+++ domain/src/main/java/com/example/domain/usecase/LoadMoviesUseCase.kt
@@ -1,16 +1,16 @@
-package com.example.topmovies.domain
+package com.example.domain.usecase
 
-import android.os.Handler
-import com.example.topmovies.models.domain.Movie
-import com.example.topmovies.repository.MoviesRepository
-import com.example.topmovies.utils.Error
-import com.example.topmovies.utils.Result
+import com.example.domain.models.Movie
+import com.example.domain.repositores.MoviesRepository
+import com.example.domain.utils.Error
+import com.example.domain.utils.HandlerWrapper
+import com.example.domain.utils.Result
 import java.util.concurrent.ExecutorService
 
 class LoadMoviesUseCase(
     private val repository: MoviesRepository,
     override val executor: ExecutorService,
-    override val handler: Handler
+    override val handler: HandlerWrapper
 ) : UseCase<Unit, List<Movie>>() {
 
     override fun execute(params: Unit): Result<Error, List<Movie>> = repository.loadNewMovies()
