+++ domain/src/main/java/com/example/domain/usecase/UpdateFavoriteMovieUseCase.kt
@@ -1,21 +1,21 @@
-package com.example.topmovies.domain
+package com.example.domain.usecase
 
-import android.os.Handler
-import com.example.topmovies.domain.UpdateFavoriteMovieUseCase.Params
-import com.example.topmovies.models.domain.Movie
-import com.example.topmovies.repository.MoviesRepository
-import com.example.topmovies.utils.Error
-import com.example.topmovies.utils.Error.ServerError
-import com.example.topmovies.utils.Result
-import com.example.topmovies.utils.Result.Failure
-import com.example.topmovies.utils.Result.Success
-import com.example.topmovies.utils.safeLet
+import com.example.domain.models.Movie
+import com.example.domain.repositores.MoviesRepository
+import com.example.domain.usecase.UpdateFavoriteMovieUseCase.Params
+import com.example.domain.utils.Error
+import com.example.domain.utils.Error.ServerError
+import com.example.domain.utils.HandlerWrapper
+import com.example.domain.utils.Result
+import com.example.domain.utils.Result.Failure
+import com.example.domain.utils.Result.Success
+import com.example.domain.utils.safeLet
 import java.util.concurrent.ExecutorService
 
 class UpdateFavoriteMovieUseCase(
     private val repository: MoviesRepository,
     override val executor: ExecutorService,
-    override val handler: Handler,
+    override val handler: HandlerWrapper,
 ) : UseCase<Params, Pair<List<Movie>, List<Movie>>>() {
 
     override fun execute(params: Params): Result<Error, Pair<List<Movie>, List<Movie>>> {
