+++ domain/src/main/java/com/example/domain/utils/Error.kt
@@ -1,4 +1,4 @@
-package com.example.topmovies.utils
+package com.example.domain.utils
 
 sealed class Error {
     object NetworkConnectionError : Error()
