+++ domain/src/main/java/com/example/domain/utils/HandlerWrapper.kt
@@ -1,0 +1,6 @@
+package com.example.domain.utils
+
+interface HandlerWrapper {
+
+    fun post(runnable: Runnable)
+}
